Updated POSIX permissions API to use short value for all method overloads instead of unsigned long, moved the assignment of `external_fa` to the same method that writes the file date attribute, and added tests for writing buffered data with and without permissions

+++ Source/UZKArchive.h
@@ -640,7 +640,7 @@ compressionMethod:(UZKCompressionMethod)method
 - (BOOL)writeData:(NSData *)data
          filePath:(NSString *)filePath
          fileDate:(nullable NSDate *)fileDate
- posixPermissions:(unsigned long)permissions
+ posixPermissions:(short)permissions
 compressionMethod:(UZKCompressionMethod)method
          password:(nullable NSString *)password
         overwrite:(BOOL)overwrite
@@ -703,7 +703,7 @@ compressionMethod:(UZKCompressionMethod)method
 - (BOOL)writeData:(NSData *)data
          filePath:(NSString *)filePath
          fileDate:(nullable NSDate *)fileDate
- posixPermissions:(unsigned long)permissions
+ posixPermissions:(short)permissions
 compressionMethod:(UZKCompressionMethod)method
          password:(nullable NSString *)password
         overwrite:(BOOL)overwrite
@@ -926,7 +926,7 @@ compressionMethod:(UZKCompressionMethod)method
  */
 - (BOOL)writeIntoBuffer:(NSString *)filePath
                fileDate:(nullable NSDate *)fileDate
-       posixPermissions:(unsigned long)permissions
+       posixPermissions:(short)permissions
       compressionMethod:(UZKCompressionMethod)method
               overwrite:(BOOL)overwrite
                     CRC:(unsigned long)preCRC
